refactor: split WorkspaceLayout into shell controller and renderer
